+++ publish.go
@@ -29,7 +29,7 @@ type Publish struct {
 }
 
 func (p *Publish) String() string {
-	return fmt.Sprintf("%s %v %v bytes",
+	return fmt.Sprintf("%s p%v %v bytes",
 		firstByte(p.fixed).String(),
 		p.packetID,
 		p.width(),