+++ cpacket.go
@@ -79,6 +79,9 @@ func (f *FixedHeader) ReadRemaining(r io.Reader) (ControlPacket, error) {
 	case UNSUBACK:
 		p = &UnsubAck{fixed: f.fixed}
 
+	case PINGREQ:
+		p = &PingReq{fixed: f.fixed}
+
 	default:
 		panic(fmt.Sprintf("ReadRemaining unhandled packet type %v", f.fixed))
 	}